refactor: generalize the Tasker bridge into a caller-agnostic profile API
The IPC surface was named for one caller (MG4Tasker), but it is just
MG4Control's external control API — list and apply driving profiles, nothing
else. Rename it so any app signed with the platform key can use it without the
naming implying a single client:

- com.mg4.control.tasker.ITaskerBridge      -> com.mg4.control.api.IProfileControl
- TaskerBridgeService                       -> ProfileControlService
- permission …TASKER_BRIDGE                 -> …CONTROL_PROFILES

No behaviour change: same two methods, same signature-level protection, same
0 km/h gate verdicts. Manifest, proguard keep rules, permission allowlist and
the verdict test move with it; French bridge comments translated to English.
